requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
cheroot==8.6.0
CherryPy==18.6.1
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
importlib-resources==5.8.0
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
more-itertools==8.13.0
mysqlclient==2.1.1
portend==3.1.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tempora==5.0.1
urllib3==1.26.10
zc.lockfile==2.0
zipp==3.8.0
